+++ src/core.c
@@ -4,47 +4,47 @@
 #include "../include/hackers_delight.h"
 #include "../include/core.h"
 
-double decimal_reciprocal(int n, int pos) {
-    return (double) iexp(10, pos - 1, n) / (double) n;
+long double decimal_reciprocal(int n, int pos) {
+    return (long double) iexp(10, pos - 1, n) / (long double) n;
 }
 
-double compute_3_4(int n, int base, int c, int (*p)(int), bool start_at_0) {
-    double sum = compute_3_4_first_sum(n, base, c, p, start_at_0) + compute_3_4_second_sum(n, base, c, p);
+long double compute_3_4(int n, int base, int c, int (*p)(int), bool start_at_0) {
+    long double sum = compute_3_4_first_sum(n, base, c, p, start_at_0) + compute_3_4_second_sum(n, base, c, p);
     
     return sum;
 }
 
 char* compute_3_4_to_base(int n, int base, int c, int (*p)(int), int places, bool start_at_0) {
-    double sum = compute_3_4(n, base, c, p, start_at_0);
+    long double sum = compute_3_4(n, base, c, p, start_at_0);
     
     return convert_floating_decimal_to_base(sum, places, base);
 }
 
-double compute_3_4_first_sum(int n, int base, int c, int (*p)(int), bool start_at_0) {
-    double sum = 0.0;
+long double compute_3_4_first_sum(int n, int base, int c, int (*p)(int), bool start_at_0) {
+    long double sum = 0.0;
     int k_start = start_at_0 ? 0 : 1;
     for (int k = k_start; k <= floor(n / c); k++) {
         int poly_result = (*p)(k);
         int num = iexp(base, n - c * k, poly_result);
         int denom = poly_result;
-        sum += (double) num / (double) denom;
+        sum += (long double) num / (long double) denom;
         sum = mod_one(sum);
     }
     
     return sum;
 }
 
-double compute_3_4_second_sum(int n, int base, int c, int (*p)(int)) {
-    double sum = 0.0;
+long double compute_3_4_second_sum(int n, int base, int c, int (*p)(int)) {
+    long double sum = 0.0;
     int k = floor(n / c) + 1;
-    double prev_sum = sum;
-    double machEps = calculate_machine_epsilon();
+    long double prev_sum = sum;
+    long double machEps = calculate_machine_epsilon();
     do {
         prev_sum = sum;
         int poly_result = (*p)(k);
-        double num = pow(base, n - c * k);
+        long double num = pow(base, n - c * k);
         int denom = poly_result;
-        sum += num / (double) denom;
+        sum += num / (long double) denom;
         sum = mod_one(sum);
         
         k++;
@@ -53,45 +53,46 @@ double compute_3_4_second_sum(int n, int base, int c, int (*p)(int)) {
     return sum;
 }
 
-double calculate_machine_epsilon(void) {
+long double calculate_machine_epsilon(void) {
     // From http://en.wikipedia.org/wiki/Machine_epsilon#Approximation_using_C
-    // Calculated Machine float epsilon:  1.19209E-07
-    // Calculated Machine double epsilon: 2.22045E-16
+    // Calculated Machine float epsilon:       1.19209E-07
+    // Calculated Machine double epsilon:      2.22045E-16
+    // Calculated Machine long double epsilon: 1.084202e-19
 
-    double machEps = 1.0;
+    long double machEps = 1.0;
  
     do {
        machEps /= 2.0;
        // If next epsilon yields 1, then break, because current
        // epsilon is the machine epsilon.
     }
-    while ((double)(1.0 + (machEps/2.0)) != 1.0);
+    while ((long double)(1.0 + (machEps/2.0)) != 1.0);
  
     return machEps;
 }
 
-double mod_one(double x) {
+long double mod_one(long double x) {
     return x - floor(x);
 }
 
-int get_int_part(double x) {
+int get_int_part(long double x) {
     return (int) x;
 }
 
 // Use  to 
-char* convert_floating_decimal_to_hex(double x, int places) {
+char* convert_floating_decimal_to_hex(long double x, int places) {
     return convert_floating_decimal_to_base(x, places, 16);
 }
 
 /**
- * Convert double to other bases.
+ * Convert long double to other bases.
  * Based on algorithm given in "Math Forum - Ask Dr. Math"
  *     http://mathforum.org/library/drmath/view/64392.html
  */
-char* convert_floating_decimal_to_base(double x, int places, int base) {
+char* convert_floating_decimal_to_base(long double x, int places, int base) {
     char* result = malloc(places * sizeof(char) + 1);
     char* pos = result;
-    double frac_part = x;
+    long double frac_part = x;
     for (int i = 0; i < places; i++) {
         frac_part = mod_one(frac_part);
         frac_part *= base;
